refactor(sec-core): move skill ledger handler
Move daemon RPC registration and request validation under handlers.
Keep the Skill Ledger job focused on scheduling and worker lifecycle.
