+++ src/agent-sec-core/tests/integration-test/skill-ledger/test_skill_ledger_daemon_integration.py
@@ -9,9 +9,12 @@ from pathlib import Path
 from typing import Any
 
 import pytest
+
 from agent_sec_cli.daemon.client import DaemonClient
-from agent_sec_cli.daemon.jobs.skill_ledger import (
+from agent_sec_cli.daemon.handlers.skill_ledger import (
     METHOD_SKILLFS_NOTIFY_CHANGE,
+)
+from agent_sec_cli.daemon.jobs.skill_ledger import (
     SKILL_LEDGER_ACTIVATION_JOB,
     SkillLedgerActivationJob,
 )
