+++ src/agent-sec-core/tests/unit-test/daemon/test_skill_ledger_activation.py
@@ -7,22 +7,15 @@ from pathlib import Path
 from typing import Any
 
 import pytest
-from agent_sec_cli.daemon.errors import BadRequestError
+
 from agent_sec_cli.daemon.jobs.skill_ledger import (
-    METHOD_SKILLFS_NOTIFY_CHANGE,
     SKILL_LEDGER_ACTIVATION_JOB,
     SkillLedgerActivationJob,
-)
-from agent_sec_cli.daemon.jobs.skill_ledger.activation import (
-    parse_skillfs_change,
-    skillfs_notify_change_handler,
 )
 from agent_sec_cli.daemon.jobs.skill_ledger.processor import (
     process_skill_change,
 )
 from agent_sec_cli.daemon.jobs.skill_ledger.protocol import SkillFsChange
-from agent_sec_cli.daemon.protocol import DaemonRequest
-from agent_sec_cli.daemon.runtime import DaemonRuntime
 from agent_sec_cli.skill_ledger.errors import UnresolvedLiveRootError
 
 
@@ -50,160 +43,6 @@ def make_skill(tmp_path: Path, name: str = "demo-skill") -> Path:
     skill_dir.mkdir()
     (skill_dir / "SKILL.md").write_text("# Demo Skill\n", encoding="utf-8")
     return skill_dir
-
-
-def request_for(skill_dir: Path, **overrides: Any) -> DaemonRequest:
-    """Build a daemon request for SkillFS notify tests."""
-    params: dict[str, Any] = {
-        "schemaVersion": 1,
-        "skillDir": str(skill_dir),
-        "skillName": skill_dir.name,
-        "eventKind": "write",
-        "paths": ["SKILL.md"],
-    }
-    params.update(overrides)
-    return DaemonRequest(
-        method=METHOD_SKILLFS_NOTIFY_CHANGE,
-        params=params,
-    )
-
-
-def test_parse_skillfs_change_validates_request(tmp_path: Path):
-    skill_dir = make_skill(tmp_path, "weather")
-
-    change = parse_skillfs_change(request_for(skill_dir).params)
-
-    assert change.skill_dir == skill_dir.resolve()
-    assert change.skill_name == "weather"
-    assert change.event_kinds == {"write"}
-    assert change.paths == {"SKILL.md"}
-
-
-def test_parse_skillfs_change_accepts_reconcile_with_empty_paths(tmp_path: Path):
-    skill_dir = make_skill(tmp_path, "weather")
-
-    change = parse_skillfs_change(
-        request_for(skill_dir, eventKind="reconcile", paths=[]).params
-    )
-
-    assert change.skill_dir == skill_dir.resolve()
-    assert change.skill_name == "weather"
-    assert change.event_kinds == {"reconcile"}
-    assert change.paths == set()
-
-
-@pytest.mark.parametrize(
-    ("overrides", "message"),
-    [
-        ({"schemaVersion": 2}, "schemaVersion"),
-        ({"skillDir": "relative-skill"}, "absolute path"),
-        ({"skillDir": "~/relative-to-home"}, "absolute path"),
-        ({"skillName": "other"}, "skillName"),
-        ({"eventKind": "chmod"}, "eventKind"),
-        ({"paths": "/absolute"}, "paths"),
-        ({"paths": ["/absolute"]}, "relative paths"),
-        ({"paths": ["../escape"]}, "relative paths"),
-        ({"paths": ["."]}, "relative paths"),
-    ],
-)
-def test_parse_skillfs_change_rejects_invalid_params(
-    tmp_path: Path,
-    overrides: dict[str, Any],
-    message: str,
-):
-    skill_dir = make_skill(tmp_path, "weather")
-
-    with pytest.raises(BadRequestError, match=message):
-        parse_skillfs_change(request_for(skill_dir, **overrides).params)
-
-
-def test_parse_skillfs_change_requires_skill_manifest(tmp_path: Path):
-    skill_dir = tmp_path / "not-a-skill"
-    skill_dir.mkdir()
-
-    with pytest.raises(BadRequestError, match="SKILL.md"):
-        parse_skillfs_change(request_for(skill_dir).params)
-
-
-def test_metadata_only_notification_is_accepted_and_ignored(tmp_path: Path):
-    skill_dir = make_skill(tmp_path, "weather")
-    runtime = DaemonRuntime(socket_path=tmp_path / "daemon.sock")
-
-    response = skillfs_notify_change_handler(
-        request_for(skill_dir, paths=[".skill-meta/latest.json"]),
-        runtime,
-    )
-
-    assert response.data["schemaVersion"] == 1
-    assert response.data["accepted"] is True
-    assert response.data["ignored"] is True
-    assert response.data["reason"] == "metadata-only change"
-
-
-def test_notify_enqueues_registered_activation_job(monkeypatch, tmp_path: Path):
-    skill_dir = make_skill(tmp_path, "weather")
-    runtime = DaemonRuntime(socket_path=tmp_path / "daemon.sock")
-    job = SkillLedgerActivationJob(
-        debounce_seconds=0,
-        worker_client=FakeWorkerClient(),
-    )
-    runtime.jobs.register(job)
-    monkeypatch.setattr(
-        "agent_sec_cli.daemon.jobs.skill_ledger.activation._resolve_managed_skill_dirs",
-        lambda: [],
-    )
-
-    async def scenario():
-        await job.start()
-        try:
-            response = skillfs_notify_change_handler(request_for(skill_dir), runtime)
-        finally:
-            await job.stop()
-        return response
-
-    response = asyncio.run(scenario())
-
-    assert response.data["schemaVersion"] == 1
-    assert response.data["accepted"] is True
-    assert response.data["ignored"] is False
-    assert response.data["queued"] is True
-    assert response.data["coalesced"] is False
-    assert response.data["skill"]["skillName"] == "weather"
-
-
-def test_notify_enqueues_reconcile_with_empty_paths(monkeypatch, tmp_path: Path):
-    skill_dir = make_skill(tmp_path, "weather")
-    runtime = DaemonRuntime(socket_path=tmp_path / "daemon.sock")
-    job = SkillLedgerActivationJob(
-        debounce_seconds=0,
-        worker_client=FakeWorkerClient(),
-    )
-    runtime.jobs.register(job)
-    monkeypatch.setattr(
-        "agent_sec_cli.daemon.jobs.skill_ledger.activation._resolve_managed_skill_dirs",
-        lambda: [],
-    )
-
-    async def scenario():
-        await job.start()
-        try:
-            response = skillfs_notify_change_handler(
-                request_for(skill_dir, eventKind="reconcile", paths=[]),
-                runtime,
-            )
-        finally:
-            await job.stop()
-        return response
-
-    response = asyncio.run(scenario())
-
-    assert response.data["schemaVersion"] == 1
-    assert response.data["accepted"] is True
-    assert response.data["ignored"] is False
-    assert response.data["queued"] is True
-    assert response.data["coalesced"] is False
-    assert response.data["skill"]["eventKinds"] == ["reconcile"]
-    assert response.data["skill"]["paths"] == []
 
 
 def test_activation_job_debounces_same_skill(monkeypatch, tmp_path: Path):
